Changes to be committed: 	modified:   .vscode/c_cpp_properties.json 	new file:   .vscode/settings.json 	new file:   .vscode/tasks.json 	modified:   AdvCRDS(RDBMS)/AdvCRDS.h 	modified:   AdvCRDS(RDBMS)/AdvCRDSImpl.cpp 	modified:   AdvCRDS(RDBMS)/ColumnValue.h 	deleted:    AdvCRDS(RDBMS)/nullptr.h
Feat: Add GetName member to ColumnValue

This commit adds a new member function, GetName, to the ColumnValue class in
the code. The GetName function retrieves the name of the column, which can be
useful for specific purposes.

Additionally, this commit addresses a clang-tidy warning regarding the 'rc'
variable in the AdvCRDSImpl.cpp file. The warning suggested a potential dead
store, but after careful review, it is clear that 'rc' is essential for the
loop condition, and the warning can be safely ignored.